Move files to handlers folder
